There are two new options on the Editor's Tileset menu: "Reload Image" and "Properties"
"Reload Image" just reloads the tileset bitmap; useful if you've got the bitmap open in a paint program.

The "Properties" option pops up a dialog that allows you to change the border of tiles on the tileset bitmap and the mask color on the bitmap.

The Properties dialog required that I change the file format and the internal representation of the tilesets a little bit.